MKPro/requirements.txt
Django==5.2.1
pillow==11.2.1